Correct component naming.
Previously descriptors `FinanceAggregateDescriptor` and `FinanceServiceDescriptor` would both have generated a default name of `finance`.  This isn't ideal as we want to distinguish and also the name should include the type.  With this change the default names will be `finance-aggregate` and `finanace-service`, respectively.
